+++ libs/datamodel/core/tests/renderer/simplification.rs
@@ -1,5 +1,4 @@
 use crate::common::parse;
-use datamodel::Datamodel;
 
 #[test]
 fn test_exclude_default_relation_names_from_rendering() {
@@ -52,23 +51,4 @@ fn test_render_relation_name_on_self_relations() {
     print!("{}", rendered);
 
     assert_eq!(rendered, input);
-}
-
-#[test]
-fn experimental_features_roundtrip() {
-    let input = r#"generator client {
-  provider             = "prisma-client-js"
-  experimentalFeatures = ["connectOrCreate", "transactionApi"]
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = "postgresql://test"
-}"#;
-
-    let dml = datamodel::parse_configuration(input).unwrap();
-    let rendered = datamodel::render_datamodel_and_config_to_string(&Datamodel::new(), &dml).unwrap();
-
-    print!("{}", rendered);
-    assert_eq!(input, rendered);
 }
